backend/requirements.txt
flask>=2.0.0
flask-cors>=3.0.10
openai>=1.12.0
python-dotenv
neomodel
neo4j
numpy
networkx
matplotlib
pandas
requests
instructor
pydantic>=2.0.0
typing-extensions
gunicorn>=20.1.0
ipykernel
ipython
chromadb